V 0.0.3.2
Wasn't happy with the way the acceleration was working so decided to modify the code. I'm pretty happy with this now. Initial acceleration rate applies when moving to a target speed which is then modified by the acceleration factor * difference in the current speed and target speed.

In short, the closer you get to the target speed, the slower the ship accelerates towards it. The further away, the faster. This will work well with sail increase as you will get an initial boost and slowly get up to speed.