V 0.0.5.3
Still WIP. Everything has been separated out. Next step is to modify HealthManager to be able to take different amounts of HealthComponent. And also receive damage from multiple different types of ammo if new one's are added.

After this, rework on the movement will be done so different movement scripts can be added in place of existing ones.

After that, connection between scripts with interfaces.