V 0.0.5.5
Few more improvements to HealthComponent and HealthManager. You can now have multiple HealthComponents take damage. E.G If hit by a round_shot it can now do 10 damage to the HullHealth and 2 to the Sails.

Also, if there is 0 damage it ignores the damage avoiding extra computation.

To also keep with SOLID principles (hopefully) DamageHandler now only references to HealthManager via it's IApplyDamage interface(will probably rename this later). So whilst it still does have a reference to it, it only has access to it through an interface.