V 0.0.8.7
Setup of attackState. Modified CannonManager and introduces some new interfaces to allow the AI to retrieve the currently loaded ammo. You can now change ammo by specific Enum type or via scrolling.

This change was needed to allow the AI to change to a specific ammo type. But to also allow the AI to retrieve data about all the ammo types in order to determine what it's combat ranges are.