V 0.0.8.7.5
Swapped around the AmmunitionData interface function. Now the CannonManager tells the AI what ammo is loaded by sending a copy of the ammodict.

My current plan with AI behaviour is if the AI is faster than the player, it will keep at range and pepper the player down. If it's slower then it'll attempt to get closer, every time all of it's cannons are loaded it'll line up a shot with whatever is in range.