V 0.0.8.7.7
More work done to the AI. Planning how the behaviours are going to work and how the AI is going to be targeting the player based on ammo type and other variables. Have also allowed the AI to know it's own speed data so it can slow down if it's more than a certain amount of degrees away from it's target to make it's turning circle smaller.