V 0.0.8.8
Did some reworking of the health script to make it more efficient. Realised every version of HealthComponent was copying data and just reacting on whatever data it was set to act on. Now they act the same but only hold the data relevant to it. So Corporeal Health only holds Corporeal receivers.

I've also added a new interface for the AI when it gets a target to interface with all the health components and add itself to a new update list. Later on when a ship goes out of range it will remove itself from this update list. This allows AI/multiple targets to get info on a ships health when it's damaged.