V 0.0.10.1
Improved AI pathfinding. Removed complete closed areas as it was causing problems with the AI. Now the AI has high cost areas near the cost.

The player and AI now also have collision without using the physics system.

Implemented smoothing on the pathfinding system to increase the number of waypoints.