V 0.12.1
Few bugs fixed. Player UI is mostly put in, cannons are missing but it's not a critical aspect. Next up is AI health bars and then work on the Game UI, including the count down, current wave and current number of enemies. Final part will be to add menus and end game screen.