1.0.1
---ADDITIONS---
Created an InputSO to hold keybindings. This will be used later on to allow the user to rebind their keys.
---FIXES---
N/A
--PROBLEMS--
N/A
---REMOVED---
N/A
---COMMENTS---
First commit of the stretch goal phase, I will be implementing any fixes/stretch goals I wanted to initially add to this project. This will probably only go on till November/December before I decide to move onto a new project.