1.0.6
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
More work on the options system, beginning to actually start creating the functionality behind the sound options so that the user can change and save the sound levels they desire.

This is being done by a scriptable object. Also expanded the keyinput scriptable object to blacklist any keys I don't want to allow being bound.