1.0.8
---ADDITIONS---
1. Options menu now effects combat and ambience sound levels in game.
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Starting to implement the sound into the actual gameplay. Currently Combat and Ambience are setup. Not the best setup I could have made but it works and since it's small and it's not going to become any larger I'm happy with the solution. Next to add in the music and move onto finishing keybindings.