1.1.3
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Started work on the newer save system. This will write to a text file to allow the user if they wish to change it. It works as a global variable as it should be accessible from everywhere and everyone should know what keys do what. This also removes the need for costly events which I will modify later.