1.2.2
---ADDITIONS---
N/A
---FIXES---
1. Using FindFirstOBjectOfType caused issues with some objects for some reason trying to find the deleted SettingsSystem.
Even though it was deleted well before they began their search... Either way, they now find it via a static declaration
of the instance..
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Polish finished, decided generics probably weren't a good idea as casting that much is bad for performance.
However it was an interesting experiment. The settings system is polished.