1.3.3
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Working on ideas to create the inventory. Setting up an inventory UI in the scene.

My idea is that there will be 1 UI, the inventory manager will hold the objects. Then when a new item is added the inventory updates. If the user clicks on the object (it will be a prefab) it'll link directly to that object's behaviour.