1.3.7
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Added a way for the upcoming store to retrieve information on what the player has in their inventory.

The store will contain a list, containing the prefabs to create or destroy objects for the store UI. It will also have a list of acceptable types. When it gets the player inventory. It'll construct both it's own and the players inventories for trade.

Once ready, these prefabs will contain the type, image, text, quantity and price. Depending on which side, they can select the object and use the buttons on the store to buy/sell.