1.3.10
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Continuing work on the store script. Have decided to use dictionaries as there will be lookups. Also, will have a selection system which will choose an item to trade, then player will then buy/sell for that item and then select another item to trade.