1.3.12
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Adding more logic to the Inventory Manager to allow the inventory manager to add new objects. But to also spawn objects added initially. This also expands into the InventoryObject itself allowing it to know it's own inventory references so it can setup it's own UI.