1.3.17
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Put in a Setup system that will allow the Inventory to determine if it's going to be setup with the player UI. The store will simply get the important aspects from it for the UI and nothing else. And behaviours won't be triggered as they require being hooked up to UI buttons.

Not as clean as I'd like but good enough.

Next step is to reconnect the InventoryManager and Store. Once that is done I'll setup the ability tracker.