1.3.18
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Wasn't happy with the layout so I've refactored the InventoryObjectSO, InventoryObject and created a new AbilityObject. InventoryObjectSO now only contains data of an object type and which scripts relate to it.

InventoryObject will only contain information about the object like price, name, it's image etc.

AbilityObject only contains information relating to an objects ability.

This will make everything far more flexible. Also due to this there will be 2 types of inventory. A PlayerInventory and a StoreInventory script. The playerInventory will create objects with abilities and the accompanying UI requirements. The StoreInventory won't. These will both inherit from a base abstract Inventory class as they will share a lot of behaviour.