1.3.20
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Was deciding if I wanted the inventory to change directly via the store or if I wanted a copy sent to the store and then only specific changes made when the user is done.

I've decided on the copy version. To ensure encapsulation is maintained so that only the Inventory can actually modify the contents of the inventory objects.

Right now I'm just fixing everything to work with this new approach, tomorrow I'll be focusing on getting the inventorymanager up and running before going back to the store to get that working.