1.3.21
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Fixing more issues. Started separating out different behaviour into the player and store inventories. Realised I don't need to pass all the data as the store and each inventory will hold the SO data. All we need to know of the object is the quantity and the ID.