1.3.24
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Turned inventory object into a new abstract object, has two inheriting classes, StoreInventoryObject and PlayerInventoryObject. Both are practically the same however StoreInventoryObject lacks UI updating and required reference in setup as it doesn't use a UI for it's stored inventory.

Also modified the StoreInventory to create a basic object when creating a new item for storage and placing it in the list. It doesn't need anything instantiated as these objects aren't going to be interactable. So they're just stored in memory.