1.3.26
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Work progressing on the InventoryManager. Also started working on the Store. It now caches all the possible UI slots it needs. Then populated them with objects it can sell/buy. The Store then only activated those objects that are currently being traded.

In future I need to add a tracking system of item IDs so they can be moved in batch to the InventoryManager to add/remove from inventory. But also add button methods to add and remove quantities in active trade.