1.3.28
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
More work on the store script. Starting to setup the buying/selling logic. I've added temp quantity values in StoreItemData so that these values can be changed. And can only be set via a method when the transaction is approved by the player.