1.3.31
---ADDITIONS---
N/A
---FIXES---
N/A
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Lots of minor changes to the store and inventory code. Instead of removing inventory items which is costly. They now are disabled and then reenabled when a new item of the same type is added.

Also, the InventoryObjects no longer handle their own adding or removing. They just set their quantity to a specific amount. The Inventory handles telling the object to set said value. Meanwhile the store and other remove components will calculate what the correct value is to set the quantity to and then pass that along to the InventoryManager.