1.3.40
---ADDITIONS---
N/A
---FIXES---
Properties can't be accessed in the editor. So to expose the InventoryObject List in Inventory I had to refactor the property into a field with a shorthand GetMethod.
---PROBLEMS---
N/A
---REMOVED---
N/A
---COMMENTS---
Setting up of the store gold prefab, also setting up the player and store with all the needed inventory script componenets.