1.3.47
---ADDITIONS---
N/A
---FIXES---
1. Removed the ID requirement in favour of just using SO's in the inspector. It's easier and if I change the ID's later down the road it means I don't have to change every instance of that ID.
2. Fixed/Improved code relating to the Slider so it now works properly allowing the user to trade for 0 items rather than forcing the player yo always buy 1.
3. Fix/Improvement allowing the player to select the same object to update/reload it.
---REMOVED---
N/A
---COMMENTS---
Everything seems to be working pretty well. The only major change that is needed is a revisit of the actual UI to indicate what the prices are etc on the tradeable objects.. As well as finishing  a few parts here and there. I do also need to actually enable the store as well when the player gets close to the port.