1.3.49
---ADDITIONS---
1. Added region which the player can sail into to open the store (not currently finished).
2. Added a script to allow text to float above the store.
---FIXES---
1. Fixed the selected item text not showing.
---REMOVED---
N/A
---COMMENTS---
Finishing off the store. Setting up the store open/close system by adding a blue area which the player can sail into and use to open the store. When they manually close the store or leave the area the store will shut.

Todo: Need to add glowing effect to the trigger region, also need to allow the floating text to always face the player.